mode blueprint to separate dir; remove nodepool from blueprint; add blueprint versioning draft
